+++ Modules/Changelog.lua
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [2.1.0] - 2026-01-03
+
+### Added
+- **CustomNames Support** - Added support for the `CustomNames` library to sync nicknames (Thanks Jods!).
+
+### Fixed
+- **Group Cache** - Fixed an issue where newly created groups were not immediately available in the cache (reported by m33shoq).
+
 ## [2.0.9] - 2026-01-03
 
 ### Added
@@ -1412,6 +1420,8 @@ local function CleanLine(line)
     line = line:gsub("/%*.-%*/", "")
     -- Remove emojis
     line = StripEmojis(line)
+    -- Remove backticks
+    line = line:gsub("`", "")
     -- Trim whitespace
     return line:gsub("^%s+", "")
 end